// src/Types/Product.ts
export interface SelectOption {
  label: string;
  id?: string;
  customOption?: boolean;
  value?: string;
}

export interface ProductFormData {
  name: string;
  slug: string;
  description: string;
  price: number;
  salePrice: number;
  sku: string;
  stock: number;
  categoryId: string;
  uniqueCategoryId?: string;
  // image: File[];
  tags: SelectOption[];
  // colorIds: SelectOption[];
  sizeIds: SelectOption[];
  materialIds: SelectOption[];
  fabricIds: SelectOption[];
  occasionIds: SelectOption[];
  colorImages?: colorImages[];
  isNewArrival: boolean;
  isBestSelling: boolean;
  isFeatured: boolean;
  showOnHomepage: boolean;
  isOffer: boolean;
  offerPrice: number;
  weight: number;
}

export interface SubAttributes {
  _id: string;
  name: string;
  colorCode?: string;
}

export interface Attributes {
  colorIds: SubAttributes[];
  sizeIds: SubAttributes[];
  materialIds: SubAttributes[];
  fabricIds: SubAttributes[];
  occasionIds: SubAttributes[];
}

export interface SEO {
  keywords: string[];
}

export interface colorImages {
  colorId: {
    colorCode: string;
    name: string;
    _id: string;
  };
  images: string[];
  _id: string;
}